[Silver III] Title: 주유소, Time: 168 ms, Memory: 15268 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -16,8 +16,8 @@ public class Main {
 
         int idx = 0;
         int min_price = price[0];
-        int result = 0;
-        int sum = 0;
+        long result = 0;
+        long sum = 0;
         while (idx < n - 1) {
             if (min_price >= price[idx + 1]) {
                 sum += dis[idx];
